Update README.md
The **Smart Attendance System** is a Python-based project designed to automate attendance tracking using facial recognition technology. It leverages the OpenCV library for image processing and the `haarcascade_frontalface_default.xml` classifier for face detection. 

**Key Components:**

1. **Student Registration (`Student's-registration.py`):** Captures images of students to create a dataset for training the facial recognition model.

2. **Model Training (`Model_Training_Script.py` and `train.py`):** Processes the collected images to train a model capable of recognizing registered students.

3. **Attendance Marking (`Attendance.py` and `main_program.py`):** Utilizes the trained model to identify students in real-time and records their attendance accordingly.

4. **Attendance Management (`view_attendance.py` and `send_it_to_email.py`):** Provides functionalities to view recorded attendance and send reports via email.

This system offers an efficient and reliable alternative to traditional attendance methods, reducing the likelihood of errors and proxy attendance. By automating the process, it saves time and ensures accurate record-keeping. 

Signed-off-by: TARUN MARSKOLHE <tarunmarskolhe14@gmail.com>